+++ src/main/java/com/kw/evcs/domain/entity/Alram.java
@@ -1,4 +1,4 @@
-package com.kw.evcs.entity;
+package com.kw.evcs.domain.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
