+++ src/main/java/com/kw/evcs/domain/repository/AlramRepository.java
@@ -1,7 +1,8 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
-import com.kw.evcs.entity.Alram;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.kw.evcs.domain.entity.Alram;
 
 public interface AlramRepository extends JpaRepository<Alram, Long> {
