+++ src/main/java/com/kw/evcs/domain/repository/ReviewChargerRepository.java
@@ -1,7 +1,8 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.kw.evcs.entity.ReviewCharger;
+
+import com.kw.evcs.domain.entity.ReviewCharger;
 
 public interface ReviewChargerRepository extends JpaRepository<ReviewCharger, Long> {
